tests/requirements/base.txt
# requirements from setup.py
djangorestframework
djangocms-text
setuptools

# other requirements
coverage
tox
